+++ src/TemplateContainer.tsx
@@ -1,52 +1,53 @@
 import * as React from 'react';
+import type { DragHandleProps, RenderTemplate } from '.';
 
 interface Props<I, C, T> {
   item: I;
-  template: new (props: any, context?: any) => T;
+  renderTemplate: RenderTemplate<I, C, T>;
   itemSelected: number;
   anySelected: number;
-  dragHandleProps: object;
-  commonProps: C;
+  dragHandleProps: Partial<DragHandleProps>;
+  commonProps?: C;
 }
 
 export default class TemplateContainer<
   I,
   C,
-  T extends React.Component<any, any>
+  T extends React.Component<Props<I, C, T>>
 > extends React.Component<Props<I, C, T>> {
-  private _template: T | undefined;
-  private readonly _templateSetter = (cmp: any) => {
-    this._template = cmp;
-  };
+  #template = React.createRef<T>();
 
-  public shouldComponentUpdate(nextProps: Props<I, C, T>): boolean {
+  shouldComponentUpdate(nextProps: Props<I, C, T>): boolean {
     return (
       this.props.anySelected !== nextProps.anySelected ||
       this.props.itemSelected !== nextProps.itemSelected ||
       this.props.item !== nextProps.item ||
-      this.props.template !== nextProps.template ||
+      this.props.renderTemplate !== nextProps.renderTemplate ||
       this.props.commonProps !== nextProps.commonProps
     );
   }
 
-  public getTemplate(): T {
-    return this._template!;
+  get template() {
+    return this.#template;
   }
 
-  public render() {
-    const { item, itemSelected, anySelected, dragHandleProps, commonProps } =
-      this.props;
-    const Template = this.props.template;
+  render() {
+    const {
+      item,
+      itemSelected,
+      anySelected,
+      dragHandleProps,
+      commonProps,
+      renderTemplate,
+    } = this.props;
 
-    return (
-      <Template
-        ref={this._templateSetter}
-        item={item}
-        itemSelected={itemSelected}
-        anySelected={anySelected}
-        dragHandleProps={dragHandleProps}
-        commonProps={commonProps}
-      />
-    );
+    return renderTemplate({
+      instanceRef: this.#template,
+      item,
+      itemSelected,
+      anySelected,
+      dragHandleProps,
+      commonProps,
+    });
   }
 }